requirements.txt
flask
FastAPI
Uvicorn
asyncio
redis
httpie
kafka-python
